Move `NeverUpToDateStep` out of `FormatterStepImpl` into its own class.
